feat(header): rename `Source Code` to `GitHub Organization`
`Source Code` is a misnomer since it doesn't navigate the reader to the full source code of Charmed HPC; it just takes you to a collection of a couple of repositories. `GitHub Organization` is better since it helps folks get connected with the project on GitHub, not necessarily dump them in a hodgepodge of repositories.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>